remove prompt_for_additional_qualifications feature flag
