+++ spec/support/test_helpers/candidate_helper.rb
@@ -173,12 +173,13 @@ module CandidateHelper
   end
 
   def candidate_fills_in_their_other_qualifications
-    fill_in t('application_form.other_qualification.qualification_type.label'), with: 'A-Level'
+    choose 'A level'
+    click_button 'Continue'
     fill_in t('application_form.other_qualification.subject.label'), with: 'Believing in the Heart of the Cards'
     fill_in t('application_form.other_qualification.institution_name.label'), with: 'Yugi College'
     fill_in t('application_form.other_qualification.grade.label'), with: 'A'
     fill_in t('application_form.other_qualification.award_year.label'), with: '2015'
-
+    choose 'No, not right now'
     click_button t('application_form.other_qualification.base.button')
     check t('application_form.other_qualification.review.completed_checkbox')
     click_button t('application_form.other_qualification.review.button')
