+++ spec/system/candidate_interface/course_selection/candidate_submitting_application_spec.rb
@@ -127,7 +127,7 @@ RSpec.feature 'Candidate submits the application' do
   end
 
   def and_i_can_see_my_other_qualification
-    expect(page).to have_content 'A-Level Believing in the Heart of the Cards'
+    expect(page).to have_content 'A level Believing in the Heart of the Cards'
     expect(page).to have_content 'Yugi College'
     expect(page).to have_content 'A'
     expect(page).to have_content '2015'
@@ -244,7 +244,7 @@ RSpec.feature 'Candidate submits the application' do
     expect(page).to have_content '07700 900 982'
     expect(page).to have_content 'Classroom Volunteer'
     expect(page).to have_content 'BA Doge'
-    expect(page).to have_content 'A-Level Believing in the Heart of the Cards'
+    expect(page).to have_content 'A level Believing in the Heart of the Cards'
     expect(page).to have_content 'I believe I would be a first-rate teacher'
     expect(page).to have_content 'Everything'
     expect(page).to have_content 'Not on a Wednesday'
